Add instructions for running the code and include a Vagrantfile
I think setting up Vagrant with VirtualBox may be a bit more complicated
than I thought it would be, so I didn't include instructions on using it
in the README.md file. I'm keeping it in there anyway since it can be a
pretty convenient way of setting up the work environment.